move files around a little
