+++ lib/embeds/index.js
@@ -1,12 +1,12 @@
-import image from './image';
-import video from './video';
-import youtube from './youtube';
-import twitter from './twitter';
+import image from './types/image';
+import video from './types/video';
+import youtube from './types/youtube';
+import twitter from './types/twitter';
 import values from 'lodash.values';
-import instagram from './instagram';
-import facebook from './facebook';
-import vine from './vine';
-import custom from './custom';
+import instagram from './types/instagram';
+import facebook from './types/facebook';
+import vine from './types/vine';
+import custom from './types/custom';
 
 export default opts => {
   const basicTypes = {
