+++ lib/embeds/types/image.js
