+++ lib/embeds/types/vine.js
